refactor: remove duplicate interface
