+++ plugin-server/src/main/ingestion-queues/session-recording-v2/sessions/blackhole-session-batch-writer.ts
@@ -1,8 +1,7 @@
 import { PassThrough } from 'stream'
 
 import { status } from '../../../../utils/status'
-import { StreamWithFinish } from './session-batch-recorder'
-import { SessionBatchWriter } from './session-batch-writer'
+import { SessionBatchWriter, StreamWithFinish } from './session-batch-writer'
 
 export class BlackholeSessionBatchWriter implements SessionBatchWriter {
     public async open(): Promise<StreamWithFinish> {
